+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - WMO textures now export correctly when exporting as GLTF.
 - Fixed issue causing cylic GLTF animations have incorrect frame ranges.
 - Fixed issue that prevented terrain exported with texture quality "None" from being importable due to referencing missing textures
+- Changing tab (Models, Textures, etc) no longer resets your scroll position.
 
 0.1.65 (31-12-2024)
 - Added option to view and export texture atlas regions from the texture viewer.